espresso_node::api::data_source

Trait StakeTableDataSource

pub(crate) trait StakeTableDataSource<T: NodeType> {
    // Required methods
    fn get_stake_table(
        &self,
        epoch: Option<EpochNumber>,
    ) -> impl Send + Future<Output = Result<Vec<PeerConfig<T>>>>;
    fn get_stake_table_current(
        &self,
    ) -> impl Send + Future<Output = Result<StakeTableWithEpochNumber<T>>>;
    fn get_da_stake_table(
        &self,
        epoch: Option<EpochNumber>,
    ) -> impl Send + Future<Output = Result<Vec<PeerConfig<T>>>>;
    fn get_da_stake_table_current(
        &self,
    ) -> impl Send + Future<Output = Result<StakeTableWithEpochNumber<T>>>;
    fn get_validators(
        &self,
        epoch: EpochNumber,
    ) -> impl Send + Future<Output = Result<IndexMap<Address, AuthenticatedValidator<BLSPubKey>>>>;
    fn get_block_reward(
        &self,
        epoch: Option<EpochNumber>,
    ) -> impl Send + Future<Output = Result<Option<RewardAmount>>>;
    fn current_proposal_participation(
        &self,
    ) -> impl Send + Future<Output = HashMap<BLSPubKey, f64>>;
    fn proposal_participation(
        &self,
        epoch: EpochNumber,
    ) -> impl Send + Future<Output = HashMap<BLSPubKey, f64>>;
    fn current_vote_participation(
        &self,
    ) -> impl Send + Future<Output = HashMap<BLSPubKey, f64>>;
    fn vote_participation(
        &self,
        epoch: EpochNumber,
    ) -> impl Send + Future<Output = HashMap<BLSPubKey, f64>>;
    fn get_all_validators(
        &self,
        epoch: EpochNumber,
        offset: u64,
        limit: u64,
    ) -> impl Send + Future<Output = Result<Vec<RegisteredValidator<PubKey>>>>;
    fn stake_table_events(
        &self,
        from_l1_block: u64,
        to_l1_block: u64,
    ) -> impl Send + Future<Output = Result<Vec<StakeTableEvent>>>;
}

Required Methods

fn get_stake_table( &self, epoch: Option<EpochNumber>, ) -> impl Send + Future<Output = Result<Vec<PeerConfig<T>>>>

Get the stake table for a given epoch

fn get_stake_table_current( &self, ) -> impl Send + Future<Output = Result<StakeTableWithEpochNumber<T>>>

Get the stake table for the current epoch if not provided

fn get_da_stake_table( &self, epoch: Option<EpochNumber>, ) -> impl Send + Future<Output = Result<Vec<PeerConfig<T>>>>

Get the DA stake table for a given epoch

fn get_da_stake_table_current( &self, ) -> impl Send + Future<Output = Result<StakeTableWithEpochNumber<T>>>

Get the DA stake table for the current epoch if not provided

fn get_validators( &self, epoch: EpochNumber, ) -> impl Send + Future<Output = Result<IndexMap<Address, AuthenticatedValidator<BLSPubKey>>>>

Get all the validators

fn get_block_reward( &self, epoch: Option<EpochNumber>, ) -> impl Send + Future<Output = Result<Option<RewardAmount>>>

fn current_proposal_participation( &self, ) -> impl Send + Future<Output = HashMap<BLSPubKey, f64>>

Get the current proposal participation.

fn proposal_participation( &self, epoch: EpochNumber, ) -> impl Send + Future<Output = HashMap<BLSPubKey, f64>>

Get the proposal participation for a given epoch.

fn current_vote_participation( &self, ) -> impl Send + Future<Output = HashMap<BLSPubKey, f64>>

Get the current vote participation.

fn vote_participation( &self, epoch: EpochNumber, ) -> impl Send + Future<Output = HashMap<BLSPubKey, f64>>

Get the vote participation for a given epoch.

fn get_all_validators( &self, epoch: EpochNumber, offset: u64, limit: u64, ) -> impl Send + Future<Output = Result<Vec<RegisteredValidator<PubKey>>>>

fn stake_table_events( &self, from_l1_block: u64, to_l1_block: u64, ) -> impl Send + Future<Output = Result<Vec<StakeTableEvent>>>

Get stake table events from L1 blocks from_l1_block..=to_l1_block.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.